requirements.txt
fastapi
psycopg2-binary
requests
uvicorn
flask
